+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 All notable changes to this project will be documented in this file.
 
+## [4.1.0] - 2025-05-16
+
+### Added
+
+### Changed
+- Bumped Version
+- Added serverUrl and pathnamePrefix to public logging methods
+- Changed setFileLoggingEnabled method names to enableFileLogging
+- Updated README.md
+
+### Fixed
+
+[4.1.0]: https://github.com/Eclypses/eclypses-aws-mte-relay-client-android/releases/tag/4.1.0
+<br><br>
+
 ## [4.0.0] - 2025-05-14
 
 ### Added
@@ -14,6 +29,7 @@ All notable changes to this project will be documented in this file.
  - General Cleanup
 
 [4.0.0]: https://github.com/Eclypses/eclypses-aws-mte-relay-client-android/releases/tag/4.0.0
+<br><br>
 
 ## [3.7.1] - 2025-05-12
 
@@ -27,7 +43,7 @@ All notable changes to this project will be documented in this file.
  - General Cleanup
 
 [3.7.1]: https://github.com/Eclypses/eclypses-aws-mte-relay-client-android/releases/tag/3.7.1
-
+<br><br>
 
 ## [3.7.0] - 2025-05-12